+++ Wrapper/VSSChunkCryptor.h
@@ -12,10 +12,10 @@ extern NSString * __nonnull const kVSSChunkCryptorErrorDomain;
 
 @interface VSSChunkCryptor : VSSBaseCryptor
 
-- (unsigned long long)startEncryptionWithPreferredChunkSize:(unsigned long long)chunkSize error:(NSError * __nullable * __nullable)error;
+- (size_t)startEncryptionWithPreferredChunkSize:(size_t)chunkSize error:(NSError * __nullable * __nullable)error;
 
-- (unsigned long long)startDecryptionWithRecipientId:(NSString * __nonnull)recipientId privateKey:(NSData * __nonnull)privateKey keyPassword:(NSString * __nullable)keyPassword error:(NSError * __nullable * __nullable)error;
-- (unsigned long long)startDecryptionWithPassword:(NSString * __nonnull)password error:(NSError * __nullable * __nullable)error;
+- (size_t)startDecryptionWithRecipientId:(NSString * __nonnull)recipientId privateKey:(NSData * __nonnull)privateKey keyPassword:(NSString * __nullable)keyPassword error:(NSError * __nullable * __nullable)error;
+- (size_t)startDecryptionWithPassword:(NSString * __nonnull)password error:(NSError * __nullable * __nullable)error;
 
 - (NSData * __nullable)processDataChunk:(NSData * __nonnull)chunk error:(NSError * __nullable * __nullable)error;
 - (BOOL)finishWithError:(NSError * __nullable * __nullable)error;
